refactor(scenarios): decompose 31-plot-plan-run-synthesis.ts under 500 lines
Decomposed the oversized 950-line scenario file into helper functions, extracting fixture setup, build functions, and git config writers into scripts/acceptance/scenarios/lib/fixture-31.ts. Reused existing type, event, and poll helpers, keeping the main scenario orchestrator under 500 lines (now 488). Lowered the file budget entry accordingly.

Closes warren-c83b
