+++ .mulch/expertise/acceptance.jsonl
@@ -36,3 +36,4 @@
 {"type":"failure","classification":"tactical","recorded_at":"2026-05-19T23:45:56.407Z","evidence":{"seeds":"warren-9e2c","commit":"688904f380d145455f4e73a2c6022fad4761f217"},"description":"Scenario 27 (warren-9e2c) expected the no-agent-commit child to keep prUrl=null via the trivial-merge branch, but warren-343a's reap.stagePlotForCommit lands a 'chore(warren): plot state' commit on every plot-bound run carrying host-side .plot/ appender writes (per-child run_dispatched, plan_run_dispatched, status_changed) back to origin. commitsAhead is therefore ≥ 1 even when the agent makes no commit, reap auto-opens a PR, and the trivial-merge contract only holds for plot-LESS projects (scenario 26).","resolution":"Update scenario 27 to assert that the no-agent-commit SEED_C run DOES open a PR (prUrl non-null) rather than stays null. Rename trivialChild/trivialRun → noAgentCommitChild/noAgentCommitRun and refresh the doc comments so the test intent matches the post-warren-343a contract. Scenario 26 still covers the plot-less trivial-merge path.","id":"mx-467a3b"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-19T23:50:24.352Z","evidence":{"seeds":"warren-73ee","commit":"c88907c76e28fc85cebf3bea9740b1ea80ffb288"},"relates_to":["mx-629c78"],"content":"Scenario 25 (scripts/acceptance/scenarios/25-plot-roundtrip.ts, warren-73ee / pl-95dd step 3) verified green end-to-end against a live in-proc warren+burrow stack with @os-eco/burrow-cli@0.3.3 (the burrow-59cd release). All four previously soft-skipped Plot assertions — PLOT_ID echo, PLOT_ACTOR echo, agent plot-append OK, reap mirror of plot.decision_made — run unconditionally and pass. Sibling scenario 27 (PlanRun + Plot) also green so the body.env parse fix did not regress the parallel plan-run path. Run command: 'bun run scripts/acceptance/run.ts --only 25'. Typical wall time ~4.6s. macOS works (in-proc only, modes: ['in-proc']).","id":"mx-0f1e6d"}
 {"type":"pattern","classification":"tactical","recorded_at":"2026-05-28T04:27:36.462Z","evidence":{"seeds":"warren-69fe","commit":"e46f2af49b7779fdaa04ec8d7b122097768bf76c"},"name":"Decomposing sequential acceptance scenarios into scenarios/lib siblings","description":"When an acceptance scenario file grows beyond the 500-line budget, it should be decomposed by extracting shared setup, assertion, and request-builder helpers into modular sibling files under scripts/acceptance/scenarios/lib/. This keeps the scenario orchestrator thin, legible, and easy to maintain while enabling reusability.","files":["scripts/acceptance/scenarios/29-plot-detail-roundtrip.ts","scripts/acceptance/scenarios/lib/types.ts","scripts/acceptance/scenarios/lib/git-helpers.ts","scripts/acceptance/scenarios/lib/poll-helpers.ts","scripts/acceptance/scenarios/lib/event-helpers.ts","scripts/acceptance/scenarios/lib/fixture-29.ts"],"id":"mx-229651"}
+{"type":"convention","classification":"tactical","recorded_at":"2026-05-28T04:34:28.519Z","evidence":{"seeds":"warren-c83b","commit":"9a0879793393298aab054685d522659899b829f2"},"content":"Decomposed scripts/acceptance/scenarios/31-plot-plan-run-synthesis.ts (950 lines) by extracting shared synthesis fixtures, build helpers, and git config redirect writers into scripts/acceptance/scenarios/lib/fixture-31.ts; leveraged existing poll-helpers and event-helpers to reduce the orchestrator scenario file under 500 lines.","id":"mx-f05190"}
