+++ scripts/acceptance/scenarios/31-plot-plan-run-synthesis.ts
@@ -39,100 +39,34 @@
  * arm runs against a real cloned project.
  */
 
-import { existsSync } from "node:fs";
-import { copyFile, mkdir, mkdtemp, readFile, writeFile } from "node:fs/promises";
+import { mkdtemp } from "node:fs/promises";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 
 import { AcceptanceError, assertEqual, assertTrue, type Scenario } from "../lib/assert.ts";
 import { WarrenHttp } from "../lib/http.ts";
 import { type BootHandle, bootInProc } from "../lib/inproc.ts";
+import {
+	fetchAllPlanRunEvents,
+	fetchAllRunEvents,
+	findTextEvent,
+	parsePlotLines,
+	readPlotEventLines,
+} from "./lib/event-helpers.ts";
+import {
+	BARE_PROJECT_URL,
+	buildBareFixture,
+	buildPlottedFixture,
+	PLOTTED_PROJECT_URL,
+	SEED_A,
+	SEED_B,
+	SEED_C,
+	writeGitConfigRedirects,
+} from "./lib/fixture-31.ts";
+import { sleep, waitForPlanState, waitForPlotStatus } from "./lib/poll-helpers.ts";
+import type { CreatePlotPlanRunResponse, ErrorEnvelope, ProjectRow } from "./lib/types.ts";
 
-interface ProjectRow {
-	readonly id: string;
-	readonly gitUrl: string;
-	readonly localPath: string;
-	readonly defaultBranch: string;
-	readonly hasSeeds?: boolean;
-	readonly hasPlot?: boolean;
-}
-
-interface PlanRunRow {
-	readonly id: string;
-	readonly planId: string;
-	readonly projectId: string;
-	readonly agentName: string;
-	readonly state: "queued" | "running" | "succeeded" | "failed" | "cancelled";
-	readonly plotId: string | null;
-}
-
-interface PlanRunChildRow {
-	readonly planRunId: string;
-	readonly seq: number;
-	readonly seedId: string;
-	readonly runId: string | null;
-	readonly state:
-		| "pending"
-		| "dispatched"
-		| "running"
-		| "pr_open"
-		| "merged"
-		| "failed"
-		| "skipped";
-}
-
-interface RunRow {
-	readonly id: string;
-	readonly state: string;
-	readonly plotId: string | null;
-}
-
-interface CreatePlotPlanRunResponse {
-	readonly planRun: PlanRunRow;
-	readonly children: readonly PlanRunChildRow[];
-	readonly synthesizedPlanId: string;
-	readonly parentSeedId: string;
-}
-
-interface PlanRunDetailResponse {
-	readonly planRun: PlanRunRow;
-	readonly children: readonly PlanRunChildRow[];
-	readonly runs: readonly RunRow[];
-}
-
-interface EventRow {
-	readonly id: number;
-	readonly runId: string;
-	readonly seq: number;
-	readonly kind: string;
-	readonly payload: Record<string, unknown> | null;
-}
-
-interface ErrorEnvelope {
-	readonly error?: { readonly code?: string; readonly message?: string };
-}
-
-interface PlotSnapshot {
-	readonly id: string;
-	readonly status: string;
-}
-
-const PLOTTED_PROJECT_URL = "https://github.com/warren-acceptance/sample-plot-plan-run-synth.git";
-const BARE_PROJECT_URL = "https://github.com/warren-acceptance/sample-plot-plan-run-synth-bare.git";
-
-// Pre-committed seeds in the plotted fixture. SEED_C is wired into
-// WARREN_STUB_NO_COMMIT_SEEDS so its dispatch drives the trivial-merge
-// branch (reap commitsAhead=0 → child → merged without GH polling).
-const SEED_A = "ah-acc31-aaaa";
-const SEED_B = "ah-acc31-bbbb";
-const SEED_C = "ah-acc31-cccc";
-const SEED_CLOSED = "ah-acc31-zzzz";
-const SD_PLAN_REF = "pl-acc31-other";
-const SEED_TS = "2026-05-18T00:00:00.000Z";
-
-const TERMINAL_PLAN_STATES = new Set(["succeeded", "failed", "cancelled"]);
 const PLAN_DEADLINE_MS = 120_000;
-const POLL_INTERVAL_MS = 500;
 const PLOT_FILE_POLL_TIMEOUT_MS = 10_000;
 /**
  * Auto-done landing budget — the plan-run state flips to `succeeded`
@@ -543,408 +477,3 @@ export const scenario: Scenario = {
 		}
 	},
 };
-
-interface BuildPlottedFixtureInput {
-	readonly fixturePath: string;
-	readonly sourceSamplePath: string;
-}
-
-interface BuildPlottedFixtureResult {
-	readonly happyPlotId: string;
-	readonly emptyPlotId: string;
-}
-
-/**
- * Build the `.seeds/`-and-`.plot/`-enabled fixture. Two Plots:
- *   - happy: 3 open seeds_issue + 1 closed seeds_issue + 1 sd_plan-
- *     shaped seeds_issue attached. Three are dispatchable after
- *     filtering; the 4th (closed) and 5th (`pl-*` ref) drop out.
- *   - empty: 1 closed seeds_issue + 1 sd_plan-shaped seeds_issue —
- *     both filtered, leaving zero candidates so the synthesis handler
- *     surfaces NoDispatchableSeedsError.
- */
-async function buildPlottedFixture(
-	input: BuildPlottedFixtureInput,
-): Promise<BuildPlottedFixtureResult> {
-	await mkdir(input.fixturePath, { recursive: true });
-	await mkdir(join(input.fixturePath, "tools"), { recursive: true });
-	await mkdir(join(input.fixturePath, ".seeds"), { recursive: true });
-
-	const burrowToml = await readFile(join(input.sourceSamplePath, "burrow.toml"), "utf8");
-	await writeFile(join(input.fixturePath, "burrow.toml"), burrowToml);
-	await copyFile(
-		join(input.sourceSamplePath, "tools", "stub-agent.sh"),
-		join(input.fixturePath, "tools", "stub-agent.sh"),
-	);
-	await copyFile(
-		join(input.sourceSamplePath, "tools", "claude-code-stub-agent.sh"),
-		join(input.fixturePath, "tools", "claude-code-stub-agent.sh"),
-	);
-	await writeFile(
-		join(input.fixturePath, "README.md"),
-		"# warren acceptance plot-plan-run synthesis fixture\n\nUsed by scripts/acceptance/scenarios/31-plot-plan-run-synthesis.ts.\n",
-	);
-
-	await writeFile(
-		join(input.fixturePath, ".seeds", "config.yaml"),
-		`project: "sample-plot-plan-run-synth"\nversion: "1"\nmax_plan_depth: 3\n`,
-	);
-	await writeFile(
-		join(input.fixturePath, ".seeds", "issues.jsonl"),
-		[
-			seedRowOpen(SEED_A),
-			seedRowOpen(SEED_B),
-			seedRowOpen(SEED_C),
-			seedRowClosed(SEED_CLOSED),
-		].join(""),
-	);
-	// Pre-seed `.seeds/plans.jsonl` empty — the synthesizer appends to
-	// it at POST time. sd plan show would reject a non-existent file,
-	// but `sd plan submit` creates it idempotently.
-	await writeFile(join(input.fixturePath, ".seeds", "plans.jsonl"), "");
-
-	const env = withGitIdentity();
-	await runIn(input.fixturePath, ["git", "init", "--initial-branch=main"], env);
-	await runIn(input.fixturePath, ["chmod", "+x", "tools/stub-agent.sh"], env);
-	await runIn(input.fixturePath, ["chmod", "+x", "tools/claude-code-stub-agent.sh"], env);
-
-	const plotEnv: Record<string, string> = { ...env, PLOT_ACTOR: "user:acceptance" };
-
-	// Happy Plot: 3 open + 1 closed + 1 sd_plan-shaped attachments.
-	await runIn(input.fixturePath, ["plot", "init", "scenario-31-happy"], plotEnv);
-	const happyList = await runIn(input.fixturePath, ["plot", "list", "--json"], plotEnv);
-	const happyPlots = JSON.parse(happyList.stdout) as ReadonlyArray<{ id: string }>;
-	const happyPlotId = happyPlots[0]?.id;
-	if (happyPlotId === undefined) {
-		throw new AcceptanceError(
-			`scenario-31 fixture: happy plot init missing id (${happyList.stdout})`,
-		);
-	}
-	await runIn(input.fixturePath, ["plot", "status", happyPlotId, "ready"], plotEnv);
-	await runIn(input.fixturePath, ["plot", "status", happyPlotId, "active"], plotEnv);
-	await runIn(
-		input.fixturePath,
-		["plot", "attach", happyPlotId, `seeds_issue:${SEED_A}`, "--role", "primary"],
-		plotEnv,
-	);
-	await runIn(
-		input.fixturePath,
-		["plot", "attach", happyPlotId, `seeds_issue:${SEED_B}`, "--role", "primary"],
-		plotEnv,
-	);
-	await runIn(
-		input.fixturePath,
-		["plot", "attach", happyPlotId, `seeds_issue:${SEED_C}`, "--role", "primary"],
-		plotEnv,
-	);
-	await runIn(
-		input.fixturePath,
-		["plot", "attach", happyPlotId, `seeds_issue:${SEED_CLOSED}`, "--role", "context"],
-		plotEnv,
-	);
-	await runIn(
-		input.fixturePath,
-		["plot", "attach", happyPlotId, `seeds_issue:${SD_PLAN_REF}`, "--role", "context"],
-		plotEnv,
-	);
-
-	// Empty-candidates Plot — both attachments will be filtered.
-	await runIn(input.fixturePath, ["plot", "init", "scenario-31-empty"], plotEnv);
-	const allList = await runIn(input.fixturePath, ["plot", "list", "--json"], plotEnv);
-	const allPlots = JSON.parse(allList.stdout) as ReadonlyArray<{ id: string }>;
-	const emptyPlotId = allPlots.map((p) => p.id).find((id) => id !== happyPlotId);
-	if (emptyPlotId === undefined) {
-		throw new AcceptanceError(
-			`scenario-31 fixture: empty plot init missing distinct id (${allList.stdout})`,
-		);
-	}
-	await runIn(input.fixturePath, ["plot", "status", emptyPlotId, "ready"], plotEnv);
-	await runIn(input.fixturePath, ["plot", "status", emptyPlotId, "active"], plotEnv);
-	await runIn(
-		input.fixturePath,
-		["plot", "attach", emptyPlotId, `seeds_issue:${SEED_CLOSED}`, "--role", "context"],
-		plotEnv,
-	);
-	await runIn(
-		input.fixturePath,
-		["plot", "attach", emptyPlotId, `seeds_issue:${SD_PLAN_REF}`, "--role", "context"],
-		plotEnv,
-	);
-
-	await runIn(input.fixturePath, ["git", "add", "."], env);
-	await runIn(
-		input.fixturePath,
-		["git", "commit", "-m", "init: plot-plan-run synthesis acceptance fixture"],
-		env,
-	);
-
-	return { happyPlotId, emptyPlotId };
-}
-
-interface BuildBareFixtureInput {
-	readonly fixturePath: string;
-	readonly sourceSamplePath: string;
-}
-
-/**
- * Bare fixture: a project clone with no `.plot/` and no `.seeds/`.
- * Used to exercise the `project_lacks_plot` arm — the
- * `hasPlot` gate fires before the seeds-cli reachability check.
- */
-async function buildBareFixture(input: BuildBareFixtureInput): Promise<void> {
-	await mkdir(input.fixturePath, { recursive: true });
-	await mkdir(join(input.fixturePath, "tools"), { recursive: true });
-
-	const burrowToml = await readFile(join(input.sourceSamplePath, "burrow.toml"), "utf8");
-	await writeFile(join(input.fixturePath, "burrow.toml"), burrowToml);
-	await copyFile(
-		join(input.sourceSamplePath, "tools", "stub-agent.sh"),
-		join(input.fixturePath, "tools", "stub-agent.sh"),
-	);
-	await copyFile(
-		join(input.sourceSamplePath, "tools", "claude-code-stub-agent.sh"),
-		join(input.fixturePath, "tools", "claude-code-stub-agent.sh"),
-	);
-	await writeFile(
-		join(input.fixturePath, "README.md"),
-		"# warren acceptance plot-plan-run synthesis bare fixture\n\nUsed by scripts/acceptance/scenarios/31-plot-plan-run-synthesis.ts (project_lacks_plot arm).\n",
-	);
-
-	const env = withGitIdentity();
-	await runIn(input.fixturePath, ["git", "init", "--initial-branch=main"], env);
-	await runIn(input.fixturePath, ["chmod", "+x", "tools/stub-agent.sh"], env);
-	await runIn(input.fixturePath, ["chmod", "+x", "tools/claude-code-stub-agent.sh"], env);
-	await runIn(input.fixturePath, ["git", "add", "."], env);
-	await runIn(input.fixturePath, ["git", "commit", "-m", "init: bare plot-plan-run fixture"], env);
-}
-
-interface GitConfigEntry {
-	readonly harnessGitConfigPath?: string;
-	readonly fakeUrl?: string;
-	readonly localPath?: string;
-}
-
-/**
- * Write a layered git-config: append the harness's existing
- * insteadOf rules (so cn/sd clones still resolve), then append
- * fresh insteadOf rules for each fixture URL → local path mapping.
- */
-async function writeGitConfigRedirects(
-	configPath: string,
-	entries: readonly GitConfigEntry[],
-): Promise<void> {
-	const out: string[] = [];
-	for (const entry of entries) {
-		if (entry.harnessGitConfigPath !== undefined) {
-			if (existsSync(entry.harnessGitConfigPath)) {
-				const body = await readFile(entry.harnessGitConfigPath, "utf8");
-				out.push(body.trimEnd());
-			}
-			continue;
-		}
-		if (entry.fakeUrl !== undefined && entry.localPath !== undefined) {
-			out.push(`[url "${entry.localPath}"]`);
-			out.push(`\tinsteadOf = ${entry.fakeUrl}`);
-		}
-	}
-	out.push("");
-	await writeFile(configPath, `${out.join("\n")}\n`);
-}
-
-function seedRowOpen(id: string): string {
-	const row = {
-		id,
-		title: `scenario-31 ${id}`,
-		status: "open",
-		type: "task",
-		priority: 3,
-		createdAt: SEED_TS,
-		updatedAt: SEED_TS,
-	};
-	return `${JSON.stringify(row)}\n`;
-}
-
-function seedRowClosed(id: string): string {
-	const row = {
-		id,
-		title: `scenario-31 ${id}`,
-		status: "closed",
-		type: "task",
-		priority: 3,
-		createdAt: SEED_TS,
-		updatedAt: SEED_TS,
-		closedAt: SEED_TS,
-	};
-	return `${JSON.stringify(row)}\n`;
-}
-
-async function waitForPlanState(
-	http: WarrenHttp,
-	planRunId: string,
-	target: string,
-	timeoutMs: number,
-): Promise<PlanRunDetailResponse> {
-	const start = Date.now();
-	let last = "unknown";
-	while (Date.now() - start < timeoutMs) {
-		const row = await http.expectJson<PlanRunDetailResponse>(
-			"GET",
-			`/plan-runs/${encodeURIComponent(planRunId)}`,
-			200,
-		);
-		last = row.planRun.state;
-		if (row.planRun.state === target) return row;
-		if (TERMINAL_PLAN_STATES.has(row.planRun.state)) {
-			throw new AcceptanceError(
-				`plan-run ${planRunId}: expected '${target}', reached terminal '${row.planRun.state}'`,
-			);
-		}
-		await sleep(POLL_INTERVAL_MS);
-	}
-	throw new AcceptanceError(
-		`plan-run ${planRunId} did not reach '${target}' within ${timeoutMs}ms (last=${last})`,
-	);
-}
-
-async function fetchAllPlanRunEvents(http: WarrenHttp, planRunId: string): Promise<EventRow[]> {
-	const events: EventRow[] = [];
-	for await (const row of http.streamNdjson(`/plan-runs/${encodeURIComponent(planRunId)}/events`)) {
-		events.push(row as EventRow);
-	}
-	return events;
-}
-
-async function fetchAllRunEvents(http: WarrenHttp, runId: string): Promise<EventRow[]> {
-	const events: EventRow[] = [];
-	for await (const row of http.streamNdjson(`/runs/${encodeURIComponent(runId)}/events`)) {
-		events.push(row as EventRow);
-	}
-	return events;
-}
-
-function findTextEvent(events: readonly EventRow[], needle: string): EventRow | undefined {
-	return events.find(
-		(e) =>
-			e.kind === "text" &&
-			typeof e.payload?.text === "string" &&
-			(e.payload.text as string).includes(needle),
-	);
-}
-
-async function readPlotEventLines(path: string): Promise<ReadonlySet<string>> {
-	const seen = new Set<string>();
-	try {
-		const body = await readFile(path, "utf8");
-		for (const line of body.split("\n")) {
-			const trimmed = line.trim();
-			if (trimmed === "") continue;
-			seen.add(trimmed);
-		}
-	} catch {
-		// File not yet present — return empty set.
-	}
-	return seen;
-}
-
-interface ParsedPlotEvent {
-	readonly type: string;
-	readonly actor: string;
-	readonly at: string;
-	readonly data: unknown;
-}
-
-function parsePlotLines(lines: ReadonlySet<string>): ParsedPlotEvent[] {
-	const out: ParsedPlotEvent[] = [];
-	for (const line of lines) {
-		let parsed: unknown;
-		try {
-			parsed = JSON.parse(line);
-		} catch {
-			continue;
-		}
-		if (typeof parsed !== "object" || parsed === null) continue;
-		const row = parsed as { type?: unknown; actor?: unknown; at?: unknown; data?: unknown };
-		if (
-			typeof row.type !== "string" ||
-			typeof row.actor !== "string" ||
-			typeof row.at !== "string"
-		) {
-			continue;
-		}
-		out.push({ type: row.type, actor: row.actor, at: row.at, data: row.data ?? null });
-	}
-	return out;
-}
-
-async function readPlotSnapshot(path: string): Promise<PlotSnapshot> {
-	const body = await readFile(path, "utf8");
-	return JSON.parse(body) as PlotSnapshot;
-}
-
-async function waitForPlotStatus(
-	path: string,
-	target: string,
-	timeoutMs: number,
-): Promise<PlotSnapshot> {
-	const start = Date.now();
-	let lastStatus = "unknown";
-	while (Date.now() - start < timeoutMs) {
-		try {
-			const snap = await readPlotSnapshot(path);
-			lastStatus = snap.status;
-			if (snap.status === target) return snap;
-		} catch {
-			// not yet present or mid-write
-		}
-		await sleep(100);
-	}
-	throw new AcceptanceError(
-		`Plot at ${path} did not reach status='${target}' within ${timeoutMs}ms (last=${lastStatus})`,
-	);
-}
-
-interface RunResult {
-	stdout: string;
-	stderr: string;
-}
-
-async function runIn(
-	cwd: string,
-	cmd: readonly string[],
-	env: Record<string, string>,
-): Promise<RunResult> {
-	const proc = Bun.spawn({
-		cmd: [...cmd],
-		cwd,
-		env,
-		stdin: "ignore",
-		stdout: "pipe",
-		stderr: "pipe",
-	});
-	const [stdout, stderr, exitCode] = await Promise.all([
-		new Response(proc.stdout).text(),
-		new Response(proc.stderr).text(),
-		proc.exited,
-	]);
-	if ((exitCode ?? 0) !== 0) {
-		throw new AcceptanceError(
-			`scenario-31 command failed (${cmd.join(" ")} in ${cwd}): exit ${exitCode}\nstderr: ${stderr}\nstdout: ${stdout}`,
-		);
-	}
-	return { stdout, stderr };
-}
-
-function withGitIdentity(): Record<string, string> {
-	return {
-		PATH: process.env.PATH ?? "",
-		HOME: process.env.HOME ?? "/tmp",
-		GIT_AUTHOR_NAME: "Warren Acceptance",
-		GIT_AUTHOR_EMAIL: "acceptance@warren.invalid",
-		GIT_COMMITTER_NAME: "Warren Acceptance",
-		GIT_COMMITTER_EMAIL: "acceptance@warren.invalid",
-	};
-}
-
-function sleep(ms: number): Promise<void> {
-	return new Promise((resolve) => setTimeout(resolve, ms));
-}
